backend/pyproject.toml
[project]
name = "drug-discovery-platform"
version = "0.1.0"
description = "End-to-end drug discovery simulation platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "celery[redis]==5.4.0",
    "redis==5.2.0",
    "httpx==0.28.0",
    "pydantic-settings==2.6.1",
    "python-multipart==0.0.12",
    "rdkit==2024.3.6",
    "anthropic==0.40.0",
    "minio==7.2.11",
    "flower==2.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aiosqlite>=0.20",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "models", "modules"]
